user_model_assignment/pyproject.toml
[tool.poetry]
name = "user-model-assignment"
version = "0.1.0"
description = ""
authors = ["ramila1 <ramilashahi2799@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.5"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
